third_party/WebKit/LayoutTests/inspector/elements/styles-4/styles-update-links-2-expected.txt

Tests that links are updated properly when editing selector.

Red text here.

Running: selectInitialNode

Running: testEditSelector


#### AFTER SELECTOR EDIT ####


[expanded] 
element.style { ()

[expanded] 
@media (min-device-width: 1px)
.should-change { (styles-update-links.css:11 -> styles-update-links.css:11:5)
    font-family: monospace;

[expanded] 
.should-change { (styles-update-links.css:7 -> styles-update-links.css:7:89)
    position: relative;
    margin: 1em;

[expanded] 
.should-change, .INSERTED-OTHER-SELECTOR { (styles-update-links.css:7 -> styles-update-links.css:7:19)
    border: 1px solid black;

[expanded] 
.should-change { (styles-update-links.css:6 -> styles-update-links.css:6:3)
    color: red;

[expanded] 
.left-intact { (styles-update-links.css:4 -> styles-update-links.css:4:1)
    padding: 1em;

[expanded] 
.left-intact { (styles-update-links.css:1 -> styles-update-links.css:1:1)
    box-sizing: border-box;

[expanded] 
.left-intact { (styles-update-links-2.css:19 -> styles-update-links-2.css:19:1)
    line-height: 1;

[expanded] 
div { (user agent stylesheet)
    display: block;

4 rule ranges are equal.
4 rule ranges are equal.
4 rule ranges are equal.
3 rule ranges are equal.
3 rule ranges are equal.
3 rule ranges are equal.
3 rule ranges are equal.
0 rule ranges are equal.
